doesn't redownload images if they already exist

+++ scraper.py
@@ -27,7 +27,8 @@ with open("cards.json", "w") as file:
         src = tr.find("th").find("img")['src']
         if str(src).endswith(".jpg"):
             filename = str(card["[Number]"]).replace("/", "-")+".jpg"
-            urllib.request.urlretrieve("http://cf-vanguard.com/en/cardlist/"+str(src), "cardfaces/"+filename)
+            if not os.path.isfile("cardfaces/"+filename):
+                urllib.request.urlretrieve("http://cf-vanguard.com/en/cardlist/"+str(src), "cardfaces/"+filename)
             card["[Image]"] = filename
 
         cards.append(card)
